+++ pyjiting/__init__.py
@@ -1,3 +1,4 @@
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 
-from .main import clear_cache, jit, jit_from_source, reg, runtime_stats
+from .main import (JITContext, clear_cache, get_llvm_ir, inspect_specializations, jit,
+                   jit_from_source, reg, runtime_stats)